+++ lib/input/s3.go
@@ -41,6 +41,9 @@ downloaded. Otherwise, the entire list of objects found when this input is
 created will be downloaded. Note that the prefix configuration is only used when
 downloading objects without SQS configured.
 
+If the download manager is enabled this can help speed up file downloads but
+results in file metadata not being copied.
+
 If your bucket is configured to send events directly to an SQS queue then you
 need to set the ` + "`sqs_body_path`" + ` field to where the object key is found
 in the payload. However, it is also common practice to send bucket events to an
@@ -68,6 +71,7 @@ This input adds the following metadata fields to each message:
 
 ` + "```" + `
 - s3_key
+- All existing file metadata (only when NOT using download manager)
 ` + "```" + `
 
 You can access these metadata fields using